Finish moving methods and helper methods into league_stats and game_stats
