+++ runner.rb
@@ -11,3 +11,4 @@ locations = {
 }
 
 stat_tracker = StatTracker.from_csv(locations)
+require "pry"; binding.pry
